+++ tests/testthat/test-helpers-data.R
@@ -195,6 +195,44 @@ test_that("rad2deg returns the correct values for 0, pi/2, 2*pi", {
   expect_equal(wrap(rad2deg(x), radians = F), rad2deg(wrap(x)))
 })
 
+test_that("make_standata() works with brmsformula", {
+  ff <- brms::bf(count ~ zAge + zBase * Trt + (1|patient))
+  sd <- make_standata(ff, data = brms::epilepsy, family = poisson())
+  expect_equal(class(sd)[1], "standata")
+})
+
+test_that("make_standata() works with formula", {
+  ff <- count ~ zAge + zBase * Trt + (1|patient)
+  sd <- make_standata(ff, data = brms::epilepsy, family = poisson())
+  expect_equal(class(sd)[1], "standata")
+})
+
+test_that("make_standata() works with bmmformula if brms >= 2.20.14", {
+  # define formula
+  ff <- bmmformula(kappa ~ 1,
+                   thetat ~ 1,
+                   thetant ~ 1)
+
+  # simulate data
+  dat <- OberauerLin_2017
+
+  # fit the model
+  if (utils::packageVersion("brms") >= "2.20.14") {
+    sd <- make_standata(formula = ff,
+                        data = dat,
+                        model = mixture3p(resp_err = "dev_rad",
+                                          nt_features = 'col_nt',
+                                          setsize = "set_size", regex = T))
+    expect_equal(class(sd)[1], "standata")
+  }
+
+  sd <- standata(ff,
+                 data = dat,
+                 model = mixture3p(resp_err = "dev_rad",
+                                   nt_features = 'col_nt',
+                                   setsize = "set_size", regex = T))
+  expect_equal(class(sd)[1], "standata")
+})
 
 test_that("get_standata() returns a standata class", {
   ff <- bmmformula(kappa ~ 1,
@@ -205,11 +243,9 @@ test_that("get_standata() returns a standata class", {
                     nt1_loc = 2,
                     nt2_loc = -1.5)
 
-  standata <- get_standata(formula = ff,
-                           data = dat,
-                           model = mixture3p(resp_err = "y" ,
-                                             nt_features = paste0('nt',1,'_loc'),
-                                             setsize = 2))
+  standata <- get_standata(ff, dat, mixture3p(resp_err = "y" ,
+                                              nt_features = paste0('nt',1,'_loc'),
+                                              setsize = 2))
   expect_equal(class(standata)[1], "standata")
 })